refactor: rename csvProcessing/ to csv/ and csvDataProcessing.ts to preprocessing.ts
