+++ components/DataUpload.tsx
@@ -2,7 +2,7 @@
 
 import FormWrapper from '@/components/FormWrapper'
 import GenericDialog from '@/components/GenericDialog'
-import { processCsvUpload } from '@/lib/csvProcessing/upload'
+import { processCsvUpload } from '@/lib/csv/upload'
 import { DataUploadEnum, FormPassbackState } from '@/lib/types'
 import { FilePlusIcon } from '@radix-ui/react-icons'
 import { Button, Flex, Text } from '@radix-ui/themes'
